CCM-7252: convert ITemplateClient to classes and no longer pass in token
