+++ lambdas/backend-api/package.json
@@ -10,7 +10,6 @@
   "dependencies": {
     "@aws-sdk/client-dynamodb": "^3.696.0",
     "@aws-sdk/lib-dynamodb": "^3.698.0",
-    "esbuild": "^0.24.0",
     "jsonwebtoken": "^9.0.2",
     "nhs-notify-backend-client": "file:../backend-client",
     "winston": "^3.17.0",
@@ -24,10 +23,10 @@
     "@types/aws-lambda": "^8.10.146",
     "@types/jest": "^29.5.14",
     "@types/jsonwebtoken": "^9.0.7",
-    "@types/uuid": "^10.0.0",
     "aws-sdk-client-mock": "^4.1.0",
     "jest": "^29.7.0",
     "jest-mock-extended": "^3.0.7",
-    "typescript": "^5.5.4"
+    "typescript": "^5.5.4",
+    "esbuild": "^0.24.0"
   }
 }
