+++ lambdas/backend-api/src/templates/api/get.ts
@@ -1,5 +1,5 @@
 import type { APIGatewayProxyHandler } from 'aws-lambda';
-import { getTemplate } from '@backend-api/templates/app/get-template';
+import { TemplateClient } from '@backend-api/templates/app/template-client';
 import { apiFailure, apiSuccess } from './responses';
 
 export const handler: APIGatewayProxyHandler = async (event) => {
@@ -11,7 +11,9 @@ export const handler: APIGatewayProxyHandler = async (event) => {
     return apiFailure(400, 'Invalid request');
   }
 
-  const { data, error } = await getTemplate(templateId, username);
+  const client = new TemplateClient(username);
+
+  const { data, error } = await client.getTemplate(templateId);
 
   if (error) {
     return apiFailure(error.code, error.message, error.details);
