+++ lambdas/backend-client/src/index.ts
@@ -2,4 +2,4 @@ export * from './types/error-cases';
 export * from './types/generated';
 export * from './types/result';
 export * from './types/template-client';
-export * from './client';
+export * from './api-client';
