+++ lambdas/backend-client/src/types/template-client.ts
@@ -2,18 +2,14 @@ import { TemplateDTO, CreateTemplate, UpdateTemplate } from './generated';
 import { Result } from './result';
 
 export interface ITemplateClient {
-  createTemplate(
-    template: CreateTemplate,
-    token: string
-  ): Promise<Result<TemplateDTO>>;
+  createTemplate(template: CreateTemplate): Promise<Result<TemplateDTO>>;
 
   updateTemplate(
     templateId: string,
-    template: UpdateTemplate,
-    token: string
+    template: UpdateTemplate
   ): Promise<Result<TemplateDTO>>;
 
-  getTemplate(templateId: string, token: string): Promise<Result<TemplateDTO>>;
+  getTemplate(templateId: string): Promise<Result<TemplateDTO>>;
 
-  listTemplates(token: string): Promise<Result<TemplateDTO[]>>;
+  listTemplates(): Promise<Result<TemplateDTO[]>>;
 }
